refactor(esm): fold the deferred naming pass into RuntimeTemplate
No separate module: the pass joins the grammar it reads on the class that
spells it, as a static beside them.

Folding the two together showed the chunk naming table had been written twice
— once by source type to reserve a stand-in, once by stand-in to spell it —
with the same template closures in both. There is one table now, read back by
stand-in, and the content hash to read is the source type it is keyed under.
