+++ lib/WebpackOptionsApply.js
@@ -524,9 +524,9 @@ class WebpackOptionsApply extends OptionsApply {
 
 		if (options.output.module) {
 			// Analyzable output is ESM output, so this is every build that can reserve a name.
-			const { fillReservedNames } = require("./esm/analyzableChunkNaming");
+			const RuntimeTemplate = require("./RuntimeTemplate");
 
-			fillReservedNames(compiler);
+			RuntimeTemplate.fillReservedNames(compiler);
 		}
 
 		if (options.experiments.syncWebAssembly) {
